refactor: create chrome-storage-client directory
