+++ src/background.ts
@@ -1,6 +1,6 @@
 import { getAll } from "./plugins/chromeAPI"
 import { ItemCollectionRepository } from "./infra/itemCollectionRepository"
-import { ChromeStorageClient } from "./infra/chromeStorageClient"
+import { ChromeStorageClient } from "./infra/chrome-storage-client/chrome-storage-client"
 
 async function main() {
   const client = new ChromeStorageClient()
