+++ src/components/importModal.tsx
@@ -3,7 +3,7 @@ import clsx from "clsx"
 import { ImportItemsAppendUsecase } from "../usecase/import-items-append-usecase"
 import { ImportItemsOverwriteUsecase } from "../usecase/import-items-overwrite-usecase"
 import { ItemCollectionRepository } from "../infra/itemCollectionRepository"
-import { ChromeStorageClient } from "../infra/chromeStorageClient"
+import { ChromeStorageClient } from "../infra/chrome-storage-client/chrome-storage-client"
 import { JSONFileClient } from "../infra/json-file-client/json-file-client"
 import { FileIOClient } from "../infra/file-io-client/file-io-client"
 import { useAlertContext } from "../contexts/alert/alertContext"
